+++ src/cli/linker.ts
@@ -30,24 +30,6 @@ export function link(asmRes: TAssemblerResult) {
 		segmentList.push([name, currPos, segObj?.length ?? 0, padLen, seg.start, seg.size]);
 	}
 
-	// if (opts.hasSegmentDirectory) {
-	// 	const lenBeforeDir = finalCode.length;
-	// 	for (const [name, offset, len, padLen, org] of segmentList) {
-	// 		const recordBuffer = makeString(null, String(name), { charSize: 1, hasLeadingLength: true });
-	// 		pushNumber(recordBuffer, { value: offset, type: 0 }, -4);
-	// 		pushNumber(recordBuffer, { value: len + padLen, type: 0 }, -4);
-	// 		pushNumber(recordBuffer, { value: org, type: 0 }, -4);
-
-	// 		const recordSize: number[] = [];
-	// 		pushNumber(recordSize, { value: recordBuffer.length + 1, type: 0 }, 1);
-
-	// 		finalCode.push(...recordSize, ...recordBuffer);
-	// 	}
-	// 	const dirOffset: number[] = [];
-	// 	pushNumber(dirOffset, { value: lenBeforeDir, type: 0 }, -4);
-	// 	finalCode.push(...dirOffset, ...makeString(null, "DISK", { charSize: 1 }));
-	// }
-
 	return {
 		finalCode,
 		dir: segmentList,